requirements.txt
numpy==2.2.3
opencv_python==4.11.0.86
pandas==2.2.3
Pillow==11.1.0
Requests==2.32.3
scikit_learn==1.6.1
selenium==4.29.0
tensorflow==2.18.0
tqdm==4.67.1
